Show buffers in vim-airline
With indexes, so that you can see what files you have open.

I tried using bufferline instead because it didn't lose me an additional
line of screen space but I couldn't get it to stop duplicating the list
of buffers in the commandline and it seems that it's possibly not
maintained anymore.