zsh: Configure aws-vault to use pass
Trying this out as an alternative to exporting env vars from `pass`
because it generates temporary sessions. I don't want to use the default
keychain backend because I trust `pass` more and the password prompt
appears within the console. Setting a prefix so that that when I add an
entry for `prod` it appears tidily in `aws-vault/prod`.