aws-vault: Use osacript for prompt on Mac
Otherwise the MFA prompt gets swallowed by the current process and it
fails to get any credentials. Using `osascript` creates a GUI prompt
that blocks the current process.

I've made it conditional so that it doesn't apply when I used my Linux
machine. I suspect it will be broken there too, but I don't care about
that yet.

This might be fixed in the future by:
https://github.com/99designs/aws-vault/issues/362