flox: Restructure activations
The `term` environment now contains `kitty` and `zsh`, called from a
Spotlight wrapper. Everything else that was in `shell` is now in
`default` which is activated in-place from RC files.

This makes it easier to pick up changes and dogfoods how we tell users
to consume default environments. I've omitted `.zprofile` though pending
the prompt issue in:

- https://github.com/flox/flox/issues/2728