git: Set email for ~/projects/flox
Using the `org/repo` structure from the `gc` alias. This won't apply to
other open source contributions outside of the Flox org or changes that
are committed on my behalf by GitHub, e.g. commit suggestions (not that
I use them) or merge commits (which we don't currently use in many
repos).